LabKey_Escritorio/requirements.txt
firebase-admin==6.5.0
pillow==10.4.0
requests==2.32.3
pyinstaller==6.11.1